pinkbombs/database/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASEID = os.getenv("DATABASEID")
PGUSER = os.getenv("PGUSER")
PGPASSWORD = os.getenv("PGPASSWORD")
PGHOST = os.getenv("PGHOST")
PGPORT = os.getenv("PGPORT")
PGDATABASE = os.getenv("PGDATABASE")
