pyproject.toml
[tool.poetry]
name = "pinkbombs"
version = "0.1.0"
description = "A dataviz app about the salmon fishing and farming industry"
authors = ["DataForGood"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.16.0"
dash-bootstrap-components = "^1.5.0"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
SQLAlchemy = "^2.0.28"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
pytest = "^8.0.2"
black = "^24.2.0"
ruff = "^0.3.0"


[tool.poetry.group.jupyter.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
